fix(ci): python full coverage line limit — split monitor completion GC tests
